backend/pyproject.toml
[project]
name = "stock-advisor-backend"
version = "0.1.0"
description = "FastAPI backend for the Stock Advisor app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.30",
    "psycopg2-binary>=2.9.10",
    "yfinance>=0.2.65",
    "langchain>=0.3.26",
    "langgraph>=0.5.3",
    "newspaper3k>=0.2.8",
    "feedparser>=6.0.11",
    "tweepy>=4.16.0",
    "requests>=2.32.4",
    "praw>=7.8.1",
    "redis>=6.2.0",
    "python-jose==3.5.0",
    "passlib==1.7.4",
]

[tool.setuptools.packages.find]
where = ["app"]
